clippy::all + targeted pedantic fixes; expand workspace allow list
- Auto-fix pass via cargo clippy --fix resolved ~50 sites (mostly
  uninlined_format_args, redundant_closure_for_method_calls, manual_let_else).
- Manual fixes for the rest: doc list-marker breakouts, manual_strip
  → strip_prefix, manual_memcpy → copy_from_slice, never_loop unrolled
  to next(), comparison_chain → match cmp, format_push_string → write!,
  inherent to_string → Display, ptr_arg signatures &Vec/&PathBuf → slices,
  cloned_ref_to_slice_refs → std::slice::from_ref, identical-arms
  if/else collapsed, cfg-gate dead constant under cuda feature, fix
  overly_complex_bool_expr `... || true` test gate.
- Expand workspace allow list with carve-outs for default_trait_access,
  fn_params_excessive_bools, struct_excessive_bools, return_self_not_must_use,
  used_underscore_binding, unnecessary_wraps, ref_option, naive_bytecount,
  float_cmp, type_complexity (algorithmic tuple types from C++ port).
- Drop elided_lifetimes_in_paths from rust lints — high noise on PyO3
  macro-generated code.

Now passes cargo clippy --workspace --all-targets -- -D warnings.
408/408 tests still pass.

+++ ferritin-align/src/ext/soialign/greedy.rs
@@ -47,7 +47,7 @@ pub fn sec2sq(
         let end = secy_bond[ju][1] as usize;
         for jj in start..end {
             let ii = invmap[jj];
-            if ii >= 0 && (i - ii) as i64 * (j as i32 - jj as i32) as i64 <= 0 {
+            if ii >= 0 && (i - ii) as i64 * (j - jj as i32) as i64 <= 0 {
                 return false;
             }
         }
